requirements.txt
# Core LLM (Groq)
groq>=0.37.0

python-dotenv
# Backend API
fastapi>=0.110.0
uvicorn>=0.27.0

# RAG framework
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0

# Vector database
chromadb>=0.5.0

# PDF loading
pypdf>=4.0.0

# Embeddings (local)
sentence-transformers>=2.6.0

# Environment variables
python-dotenv>=1.0.0




